requirements.txt
torch
torchvision
numpy
pillow
customtkinter
streamlit
matplotlib